requirements.txt
Django==1.8.6
django-markdown==0.8.4
djangorestframework==3.3.1
Markdown==2.6.4